fix: move deprecated code to separate package
also try to fix concert break with original tests
